third_party/blink/web_tests/http/tests/security/mixedContent/insecure-formSubmission-in-main-frame-allowed-expected.txt

CONSOLE WARNING: Mixed Content: The page at 'https://127.0.0.1:8443/security/mixedContent/resources/frame-with-insecure-formSubmission.html' was loaded over a secure connection, but contains a form that targets an insecure endpoint 'http://example.test:8080/security/resources/boring.html'. This endpoint should be made available over a secure connection.
This test opens a window that shows a form with "action" pointing to insecure location. We should trigger a mixed content callback even though we've set the preference to block this, because we've overriden the preferences via a web permission client callback.
